+++ internal/server/remote.go
@@ -405,6 +405,8 @@ func (r *remotePlayer) emitEv(ev any, send func(byte, any)) {
 		send(attachproto.MsgBundleOpen, attachproto.BundleMark{})
 	case bundleClose:
 		send(attachproto.MsgBundleClose, attachproto.BundleMark{})
+	case attachproto.PostEffects:
+		send(attachproto.MsgPostEffects, ev)
 	case attachproto.GhostRecipe:
 		send(attachproto.MsgGhostRecipe, ev)
 	case attachproto.TransientBlock: